--- .eslintrc.json
@@ -1,22 +1,0 @@
-{
-  "root": true,
-  "env": {
-    "node": true,
-    "jest": true,
-    "es2021": true
-  },
-  "parserOptions": {
-    "ecmaVersion": "latest",
-    "sourceType": "module"
-  },  
-  "extends": [
-    "eslint:recommended"
-  ],
-  "rules": {
-    "quotes": ["error", "single", { "avoidEscape": true }],
-    "semi":   ["error", "always", { "omitLastInOneLineBlock": true}],
-    "no-console": "off",
-    "indent": ["error", 4],
-    "no-trailing-spaces": ["error"]
-  }
-}
