+++ src/index.js
@@ -1,5 +1,3 @@
-#!/usr/bin/env node
-
 import chalk from 'chalk';
 import { Command } from 'commander';
