+++ src/lib/include-file-manager.js
