+++ src/lib/utils/fs-utils.js
