+++ src/lib/utils/sha1-stream.js
