+++ src/lib/watcher.js
